10 64-bit binaries busted Reviewed by: estseg@gmail.com Approved by: gdamore@nexenta.com

+++ usr/src/lib/libc/port/locale/_ctype.h
@@ -83,6 +83,6 @@
 #define	_CTYPE_SWM	0xe0000000U		/* Mask for screen width data */
 #define	_CTYPE_SWS	30			/* Bits to shift to get width */
 
-unsigned long	___runetype(int);
+unsigned int	___runetype(int);
 
 #endif /* !__CTYPE_H_ */
